+++ index.html
@@ -7,7 +7,10 @@
     <link rel="stylesheet" href="./src/styles/fonts.css" />
     <link rel="stylesheet" href="./src/styles/global.css" />
     <link rel="stylesheet" href="./src/components/header/header.css" />
-    <link rel="stylesheet" href="./src/components/transactionform.css" />
+    <link
+      rel="stylesheet"
+      href="./src/components/transaction-form/transactionform.css"
+    />
     <link rel="stylesheet" href="./src/components/transactionList.css" />
     <title>Document</title>
   </head>
