--- src/components/TransactionForm.js
@@ -1,61 +1,0 @@
-import { renderTransactionForm } from "./TransactionFormView.js";
-import DateState from "../store/DateState.js";
-import {
-  CategorySelectObserver,
-  DateInputObserver,
-  DescriptionLengthObserver,
-  SubmitButtonObserver,
-  TransactionFormObserver,
-  TransactionTypeButtonObserver,
-} from "../observers/transactionFormObservers.js";
-import FormState from "../store/FormState.js";
-import itemsState from "../store/ItemsState.js";
-
-export function initTransactionForm() {
-  renderTransactionForm();
-
-  new DateInputObserver(DateState);
-  new TransactionTypeButtonObserver(FormState);
-  new DescriptionLengthObserver(FormState);
-  new CategorySelectObserver(FormState);
-  new SubmitButtonObserver(FormState);
-  new TransactionFormObserver(FormState);
-
-  const form = document.getElementById("transaction-form");
-  if (!form) return;
-
-  form.addEventListener("input", (e) => {
-    const { name, value } = e.target;
-
-    if (!name) return;
-    FormState.setFormState({ [name]: value });
-  });
-
-  form.addEventListener("click", (e) => {
-    const toggleBtn = e.target.closest("#transaction-form-toggle-btn");
-    if (!toggleBtn) return;
-
-    const currentType = FormState.getFormState().type;
-    const newType = currentType === "withdraw" ? "deposit" : "withdraw";
-    FormState.setFormState({ type: newType, category: "" });
-  });
-
-  // submit 이벤트 위임 등록
-  form.addEventListener("submit", (e) => {
-    e.preventDefault();
-    const formState = FormState.getFormState();
-    const newItem = {
-      ...formState,
-      id: formState.editId ?? Date.now(),
-    };
-
-    const updatedItems = formState.editId
-      ? itemsState
-          .getItems()
-          .map((item) => (item.id === formState.editId ? newItem : item))
-      : [...itemsState.getItems(), newItem];
-
-    itemsState.setItems(updatedItems);
-    FormState.resetFormState();
-  });
-}
